fix denomination of widget
