+++ lib/src/all_screens.dart
@@ -1,7 +1,6 @@
 import 'package:featlink_app/src/screens/_create_profile/create_profile_step_110_screen.dart';
-import 'package:featlink_app/src/screens/page_d_day_zone_screen/page_d_day_zone_screen.dart';
+import 'package:featlink_app/src/screens/d_day_zone_screen/d_day_zone_screen.dart';
 import 'package:flutter/material.dart' show Widget;
-
 import 'screens/_create_profile/create_profile_step_1o2_screen.dart';
 import 'screens/_create_profile/create_profile_step_1o5_screen.dart';
 import 'screens/_create_profile/create_profile_step_2o2_screen/create_profile_step_2o2_screen.dart';
@@ -62,7 +61,7 @@ final Map<String, Widget Function()> allScreens = {
 
   'notifications_screen.dart': () => const NotificationsScreen(),
   'filter_screen.dart': () => const FilterScreen(),
-  'page_d_day_zone_screen.dart': () => const PageDdayZoneScreen(),
+  'd_day_zone_screen.dart': () => const DDayZoneScreen(),
 
   // Ajoutez d'autres chemins de fichiers ici
 };
