+++ lib/src/screens/d_day_zone_screen/widgets/the_d_day_tab.dart
@@ -1,6 +1,6 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:featlink_app/generated/locale_keys.g.dart';
-import 'package:featlink_app/src/components/ddaycard.dart';
+import 'package:featlink_app/src/components/d_day_card.dart';
 import 'package:featlink_app/src/config/app_colors.dart';
 import 'package:flutter/material.dart';
